feat: verify if admin before performing delete operation

+++ resources/item.py
@@ -2,7 +2,7 @@
 
 # from flask import request
 from flask_smorest import abort, Blueprint
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from flask.views import MethodView
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 # from uuid import uuid4
@@ -25,6 +25,10 @@ class Store(MethodView):
     
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege reqiuired.")
+
         item = ItemModel.query.get_or_404(item_id)
         try:
             db.session.delete(item)
